Test with Java 21 (#771)
* Remove unnecessary dependency declaration

Already declared in the parent pom, no need to declare here.

* Use most recent plugin BOM - 2496.vddfca_753db_80

* Use most recent parent pom - 4.74

* Test with Java 21

Java 21 released Sep 19, 2023. We'd like to announce full support for
Java 21 in early October and would like the most used plugins to be
compiling and testing with Java 21.

The acceptance test harness and plugin bill of materials tests are already
passing with Java 21. This is a further step to improve plugin readiness
for use with Java 21 and for development with Java 21.

Java 11 will be unsupported by Eclipse Temurin and some other Java
providers in October 2024. We'll need to move past Java 11 by that
time. It does not change the supported Java version or the byte code
that is being generated.

Not testing Java 11 configuration explicitly because the Java 11 byte
code is generated by the Java 17 and Java 21 compilers and Java 11 is
still being tested in the plugin bill of materials on LTS releases.